// src/api/plugins/sandbox/models.gen.ts
export type CommandLogStream = "stdout" | "stderr";
// ...
export interface GetCommandData {
  /**
   * The UUID of the command.
   */
  uuid: string;
  /**
   * The command line the command was started with.
   */
  cmdline: string;
  /**
   * The current working directory, if supplied at startup, else null
   */
  cwd: string | null;
  /**
   * Environment variables, if supplied at startup, else null
   */
  env: Record<string, string> | null;
  /**
   * The exit code of the command. It is `null` while the command is still
   * running.
   */
  exitcode: number | null;
}